backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Rate Limiting
slowapi>=0.1.9

# HTTP Client
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
anyio>=4.2.0

# Logging
python-json-logger>=2.0.7

# Code Quality
ruff>=0.1.14
black>=24.1.0
mypy>=1.8.0

# Pre-commit
pre-commit>=3.6.0

# Supabase (for Phase 2)
supabase>=2.3.4
